add dayjs in prep to remove moment
We need to update all the filters from vue 2, like: `{{ myDate | moment('YYYY') }}`, since they're not supported in vue3.

Vue-Moment hasn't been updated in 4 years and momentjs is at end of life. Now is the time to swap moment for dayjs, which has basically the same api as moment but is maintained and more lightweight.