fix date shifting
This resolves an issue where the frontend would submit a ISO date like `2024-01-01T08:00:00` but api would return `2024-01-01T00:00:00` (presumably since the db column is just date?). Rather than trying to fuss with parsing dates and transforming them depending on server/client, this just changes the model to always return the YYYY-MM-DD iso string, without any time.

I explored using a resource to transform the response too, but we'd also need to do a transform on a user resource... so this seemed cleaner, but maybe a bit more magical.