Fix: autocomplete clicking in safari
Background:
When typing in the person search field, the dropdown appears with a list of buttons that are choices to complete the user.

When there's no input focus or there's no focus within (say on a button), the menu should disappear using the css:
```
.person-search:not(:focus-within) .person-search__results {
  display: none;
}
```

However, Safari behaves differently. A button is not considered a focusable element in Safari. So clicking on the button, causes the person-search NOT to have focus within it, which hides the potential results and thus prevents the button from being clicked. (Or at least that's what I think is happening).

I tried giving the buttons a tabindex to explicitly make them focusable, but that didn't work.

So, this reworks the component to use JS to open/close the autocomplete box.

In Chrome and Firefox, this is fine.