normalize course records
When getting courses by terms, it's possible for there to be more than one course record in the set since there could be more than 1 TA or primary instructor  (or both).

This changes things so that a `course` contains an `instructors` list which has all the course instructors. Now, each course within a term should be unique, and all the course's instructors or ta's will be on the single record.

This should make it easier to pivot from listing records by instructors or ta's to by courses.