fix tests: change to useTermsStore to init on use
Tests are failing when we initialize the `termsStore` in `App.vue`. Because Bandaid server data isn't available, a 500 error will show up in the console, which causes Cypress to fail.

Instead of initializing the `termStore` in `App`, this commit refactors to initialize  only when `useTermsStore` is invoked.

Mocking the termsStore was another option, but we're currently not testing anything in bluesheet that relies in bandaid.