start moving vuex user store to pinia
managing leave artifact state is getting complicated, and there's a lot of prop drilling. Leave state is on the user object, which has it's sorce of truth in a `vuex` store.

We deferred migrating from vuex to pinia, but now seems like the time to complete this migration if we want to use the store to help with leave state.

This starts the migration from vuex to pinia. We also migrate the errors to an error store, and bring over some components from elevator to display errors to the user.