⚠️ Make emplid et all mass assignable
It would be possible for a db user to have a null emplid (since it was created using `$ldapUser->toArray()` but `emplid` isn't mass assignable, resulting in an empty `emplid`.

This would cause issues when trying to match up dept info from bandaid with user data generated from LDAP.

On `develop` branch, the UserService just gives up and returns a null user. In this branch, we chage to `updateOrCreate` and use `umndid` to update/create since it's required to be unique in the schema.