revert n+1 query fix
probably out of scope, certainly untested

fix n+1 query - remove load('officialGroupType')

--
Not sure what the intent is here, but it seems like it might be unnecessary (routes that I've been looking at already have this info). Maybe not tho?

Whatever the case, this loading prob should happen in one batch in the controller, not the resource to avoid n+1 queries