fix mismatch ldap/bandaid emplids
`updateOrCreate` uses `tap` under the hood and so returns the values passed to it rather than the actual user from the database.

Since LDAP's emplid is a string (possibly with a leading 0), but Bluesheet stores emplid's as an int, we need to cast the ldap emplid as an int so that the right values are returned.