add cached version of ldap lookup user and use it in UserService
`findOrCreateManyByEmplId` looks for a user in the db, and if it doesn't exist it searches ldap so that the user can be created.

There are a number of emplids that don't have entries in LDAP. (In Psych, for example, there's dozens.) Which means, each time we `getDeptEmployees` there will be dozens of ldap lookups.

Presumably LDAP data doesn't change often. This adds a cached version of the LDAP lookup `lookupUserCached`, so that we avoid looking for missing emplids.

TIL that Larave's `Cache::remember` method will always interpret a null value as a cache miss. So, when doing a `lookupUserCached` , we shove the ldap response in an assoc. array like: `['data' => null]`, which will be interpreted as SOME value by remember.

I suppose I could have used get and put caching... but I like the affordances of giving remember a function.