just track active item
tracking lists in this store makes for 2 sources of truth: whatever's behind the `list` prop (in this case the `CourseSectionStore`) and the `DragStore` . Reconciling gets complication, and requires local state.

instead, lets just track the active item, source list and target list ids.

This also fixes the typescript complains about using generics with pinia. To resolve the issue, we change useDragDrop under the hood, so it's basically a factory for stores namespaced by group names.