add Pest for api testing
The goal of this feature branch is to let users create courses that don't exist yet in the SIS. In my first experiments, things would break – but it would be difficult to discover when and where until later because currently there are no tests for the course planning section.

I've been avoiding tests because the course planning section depends on data from Bandaid. To add tests, we'll need to mock Bandaid's API responsel. And since it's an API call via Laravel, cypress testing won't intercept... which requires writing tests for php.

Since we have no real PHP unit tests, we swap PHP unit for Pest: it's simpler and less verbose. Also, it has basically the same api as other frontend testing frameworks.

The legacy Dusk tests are removed since Cypress is handling e2e testing now.