Layout for Leave Artifacts broken
Fixes #106

Removes the extra <tbody> wrapping the main contents.

We're using <tbody> to wrap the leave artifact details section. (There was probably a better way to go about this layout, but I wanted the details columns to align with the main artifact columns.)

Whatever the case, nested <tbody> is causing the problem.

Tables can have multiple <tbody>, but I guess nested <tbody> is invalid. To resolve, this removes the main wrapping tbody.