add jest for testing
Going to transform the table into json to get the format we want, rather than munging the table DOM.

We'll use `usePersonTableData` composable to do the work. But, I want to test it to make sure it's not messing things up.

If this works, we could probably eliminate a bunch of logic within the PersonTable components and just use the records from `usePersonTableData`