arrrrrrrrgggg - 'viewAny' policy not invokes because user is superadmin
Spent WAY too long trying to figure out why my `viewAny` policy isn't invoked at all on some tests! It's because the user who owns the leave is an admin -- and I was logging in as such.

There's a gate in the AuthServiceProvider which skips any auth checks for super admins