handle case where leaves data is not returned for user
when viewing a user, the currentUser may not have have permission to view another user's leaves. This causes userStore to throw an error and leaves a blank page for currentUser - even though they may be able to view other data.

This handles an undefined leaves value from the server.  An empty leaves array will be appended to the user so that all user data has the same shape.

To handle leave viewing, we explicitly check whether the user can view/edit leaves or is current user. If not, no table will be rendered.