remove unused batch update user leaves endopint
This removes the `PUT /api/user/{leaveOwner}/leaves` and supporting methods and tests , which was previously used to do a batch update of all of a user's leaves.

Formerly, the UI edited leaves similarly to how groups edited members.

Then, at some point, the UI changed so that leaves could be edited individually via `PUT /api/leaves/:id`, making the batch edit endpoint obsolete.

This change is related to scoping group admin permissions to a particular group. (There's no need to add a policy or update tests around this endpoint.)