fix: cancelling new leave create leaves row in local store
When a user clicks "Add Leave" and then cancel, the leave will persist in the local store – showing an extra row to the users. (No actual new leave is created.)

This fixes the issue, removing the leave from the local store when the Cancel button is clicked.

Adds a test.