Fix weird gantt chart header formatting
When viewing the gantt chart, the Name column header shows on a separate (not full) row as the dates.

Previously the table header elements were rendered in a separate component as the member list and gantt chart.

This makes the Gantt and MemberList components slightly more independent, encapsulating the full table and their respective headers.

Sorting props are passed down to Gannt and MemberList, and then when a table header is clicked, a sort event is emitted.

Sortable Link is changed to a button element for accessibility.

Not all table headers are sortable – specifically in the Gantt, the date labels aren't, thought they look like they might be. So, a subtle sort icon is added to SortableLink to distinguish sortable items from non-sortable.