fix: null returned from getTerms()
In previous PR, we changed the allow_redirects option to false to fix a failing CI test.
```
Http::globalOptions([
  'allow_redirects' => false,
]);
```

The side effect was that `Bandaid::getTerms()` would fail because it relied on a redirect. It's url `/classes/terms/` contained a trailing slashes, which previously would be redirected to `/classes/terms` (no trailing slash).

This removes the trailing slash to resolve the issue.