fix: moving sections causes all sorts of bad
when moving sections in course planning mode, they would sometimes move properly and other times they would just disappear or move multiple sections

This was because the id of a section (within the pinia store) is no longer stable after updating api keys.

e.g. before course section would have id of `db-4` and even if the section would move, it still would be `db-4`.

after updating, the section id is based on the course id, section, and term – and so it's no longer stable in the store when changing from one term to another.

To resolve the issue, this removes the old section id and uses the new updated section id.

We could have also keyed by dbID for local sections, but I think it's better be be consistent between the sis section and the local db sections.