fix: disappearing courses when changed
when a planned section/enrollment is changed, the id of that section might also change causing issues in the interface.

This keeps the id stable on updates, so that the store entry (keyed by id) can be updated, rather than deleted and replaced.

An alternative approach would be to update the store and make sure old values are completely deleted and new values are added with the new id -- which might be better, but require more changes

Fixes #211