add create new option button
CreateGroup modal currently allows new options to be created when a user hits enter, so maybe it's better for the "Add New Option" button to be built into the component, and then we can handle a click and enter event the same way.

If we just used a slotted component, we'd need the parent to be responsible for closing the options menu of the child... which isn't awful, but might be overly complex.